Move save field from Scenario to Stage model
