+++ tests/models/test_scenario.py
@@ -58,178 +58,16 @@ def test_scenario_with_complex_stages():
     assert scenario.stages[2].data is True
 
 
-def test_scenario_with_valid_save_field():
+def test_scenario_with_stages_containing_save_field():
     data = {
-        "stages": [{"name": "test", "data": "data"}],
-        "save": {"user_id": "user.id", "user_name": "user.name", "first_item": "items[0]", "_private_var": "data._private", "complex_path": "users[*].profile.name"},
+        "stages": [
+            {"name": "test", "data": "data", "save": {"user_id": "user.id", "user_name": "user.name"}},
+            {"name": "test2", "data": "data2"}
+        ]
     }
     scenario = Scenario.model_validate(data)
-    assert scenario.save is not None
-    assert scenario.save["user_id"] == "user.id"
-    assert scenario.save["user_name"] == "user.name"
-    assert scenario.save["first_item"] == "items[0]"
-    assert scenario.save["_private_var"] == "data._private"
-    assert scenario.save["complex_path"] == "users[*].profile.name"
-
-
-@pytest.mark.parametrize("save_value,expected", [
-    (None, None),
-    ({}, {}),
-])
-def test_scenario_save_field_optional_states(save_value, expected):
-    data = {"stages": [{"name": "test", "data": "data"}], "save": save_value}
-    scenario = Scenario.model_validate(data)
-    assert scenario.save == expected
-
-
-def test_scenario_without_save_field():
-    data = {"stages": [{"name": "test", "data": "data"}]}
-    scenario = Scenario.model_validate(data)
-    assert scenario.save is None
-
-
-@pytest.mark.parametrize("invalid_key,expected_error", [
-    ("1invalid", "'1invalid' is not a valid Python variable name"),
-    ("user-id", "'user-id' is not a valid Python variable name"),
-    ("user id", "'user id' is not a valid Python variable name"),
-    ("user@id", "'user@id' is not a valid Python variable name"),
-    ("", "'' is not a valid Python variable name"),
-])
-def test_scenario_save_invalid_python_variable_names(invalid_key, expected_error):
-    data = {"stages": [{"name": "test", "data": "data"}], "save": {invalid_key: "user.id"}}
-    with pytest.raises(ValidationError) as exc_info:
-        Scenario.model_validate(data)
-    assert expected_error in str(exc_info.value)
-
-
-@pytest.mark.parametrize("invalid_jmespath", [
-    "user.[invalid}",
-    "user.",
-    "users[0",
-])
-def test_scenario_save_invalid_jmespath_expressions(invalid_jmespath):
-    data = {"stages": [{"name": "test", "data": "data"}], "save": {"user_id": invalid_jmespath}}
-    with pytest.raises(ValidationError) as exc_info:
-        Scenario.model_validate(data)
-    assert "is not a valid JMESPath expression" in str(exc_info.value)
-
-
-@pytest.mark.parametrize("keyword", [
-    "class", "for", "if", "else", "while", "def", "return", "try", "except", "import", "from", "as",
-    "match", "case", "_", "type"
-])
-def test_scenario_save_invalid_keywords(keyword):
-    data = {
-        "stages": [{"name": "test", "data": "data"}],
-        "save": {keyword: "user.id"}
-    }
-    with pytest.raises(ValidationError) as exc_info:
-        Scenario.model_validate(data)
-    assert f"'{keyword}' is a Python keyword and cannot be used as a variable name" in str(exc_info.value)
-
-
-def test_scenario_save_valid_underscore_variable_names():
-    data = {
-        "stages": [{"name": "test", "data": "data"}],
-        "save": {
-            "__": "user.double_underscore",
-            "___": "user.triple_underscore",
-            "_private": "user.private",
-            "__private__": "user.dunder_private"
-        }
-    }
-    scenario = Scenario.model_validate(data)
-    assert scenario.save["__"] == "user.double_underscore"
-    assert scenario.save["___"] == "user.triple_underscore"
-    assert scenario.save["_private"] == "user.private"
-    assert scenario.save["__private__"] == "user.dunder_private"
-
-
-def test_scenario_save_valid_complex_jmespath_expressions():
-    data = {
-        "stages": [{"name": "test", "data": "data"}],
-        "save": {
-            "filtered_users": "users[?age > `18`]",
-            "mapped_names": "users[*].name",
-            "first_active": "users[?active][0]",
-            "nested_access": "data.nested.deeply.nested.value",
-            "pipe_expression": "users | [0]",
-            "function_call": "length(users)",
-            "conditional": "users[0] || `default`"
-        }
-    }
-    scenario = Scenario.model_validate(data)
-    assert scenario.save["filtered_users"] == "users[?age > `18`]"
-    assert scenario.save["mapped_names"] == "users[*].name"
-    assert scenario.save["first_active"] == "users[?active][0]"
-    assert scenario.save["nested_access"] == "data.nested.deeply.nested.value"
-    assert scenario.save["pipe_expression"] == "users | [0]"
-    assert scenario.save["function_call"] == "length(users)"
-    assert scenario.save["conditional"] == "users[0] || `default`"
-
-
-def test_scenario_save_multiple_validation_errors():
-    data = {
-        "stages": [{"name": "test", "data": "data"}],
-        "save": {
-            "1invalid": "user.id",
-            "valid_name": "user.[invalid}"
-        }
-    }
-    with pytest.raises(ValidationError) as exc_info:
-        Scenario.model_validate(data)
-    assert "'1invalid' is not a valid Python variable name" in str(exc_info.value)
-
-
-def test_scenario_save_keyword_vs_invalid_variable_error_precedence():
-    data = {
-        "stages": [{"name": "test", "data": "data"}],
-        "save": {"1class": "user.id"}
-    }
-    with pytest.raises(ValidationError) as exc_info:
-        Scenario.model_validate(data)
-    assert "'1class' is not a valid Python variable name" in str(exc_info.value)
-
-
-def test_scenario_save_valid_non_keyword_identifiers():
-    data = {
-        "stages": [{"name": "test", "data": "data"}],
-        "save": {
-            "user_id": "user.id",
-            "userName": "user.name",
-            "USER_NAME": "user.name",
-            "user123": "user.id",
-            "_user": "user.private",
-            "__internal__": "user.internal",
-            "MyClass": "user.class_name",
-            "for_user": "user.for_field",
-        }
-    }
-    scenario = Scenario.model_validate(data)
-    assert len(scenario.save) == 8
-    assert scenario.save["user_id"] == "user.id"
-    assert scenario.save["MyClass"] == "user.class_name"
-    assert scenario.save["for_user"] == "user.for_field"
-
-
-def test_individual_annotated_types():
-    from pydantic import BaseModel
-
-    from pytest_http.models import JMESPathExpression, ValidPythonVariableName
-
-    class TestModel(BaseModel):
-        var_name: ValidPythonVariableName
-        jmes_expr: JMESPathExpression
-
-    valid_data = {"var_name": "valid_name", "jmes_expr": "user.id"}
-    model = TestModel.model_validate(valid_data)
-    assert model.var_name == "valid_name"
-    assert model.jmes_expr == "user.id"
-
-    with pytest.raises(ValidationError) as exc_info:
-        TestModel.model_validate({"var_name": "1invalid", "jmes_expr": "user.id"})
-    assert "'1invalid' is not a valid Python variable name" in str(exc_info.value)
-
-    with pytest.raises(ValidationError) as exc_info:
-        TestModel.model_validate({"var_name": "valid_name", "jmes_expr": "user.[invalid}"})
-    assert "is not a valid JMESPath expression" in str(exc_info.value)
+    assert len(scenario.stages) == 2
+    assert scenario.stages[0].save is not None
+    assert scenario.stages[0].save["user_id"] == "user.id"
+    assert scenario.stages[0].save["user_name"] == "user.name"
+    assert scenario.stages[1].save is None
